requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.17
anthropic==0.40.0
pyodbc==5.2.0
openpyxl==3.1.5
python-dotenv==1.0.1
